requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
python-dotenv
seaborn
matplotlib
fastapi
uvicorn
requests
pytest
httpx
